Refactored code to split per tab
