+++ lib/api/driver_components.dart
@@ -21,7 +21,7 @@ import 'package:boxbox/helpers/driver_image.dart';
 import 'package:boxbox/helpers/loading_indicator_util.dart';
 import 'package:boxbox/helpers/request_error.dart';
 import 'package:boxbox/helpers/team_background_color.dart';
-import 'package:boxbox/Screens/driver_details.dart';
+import 'package:boxbox/Screens/DriverDetails/details.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
